+++ src/renderer/pages/ProfileEditor.tsx
@@ -42,12 +42,8 @@ export default function ProfileEditor() {
   }
 
   const ModButton = ({ name }: { name: string }) => {
-    const [isHovered, setIsHovered] = useState(false)
-
     return (
-      <div
-        onMouseEnter={() => setIsHovered(true)}
-        onMouseLeave={() => setIsHovered(false)}
+      <div className="m-[-3px] border-[3px] border-[#1E1E1F]"
         onClick={() => {
           if (profileRuntime === 'Vanilla') {
             alert('Cannot add mods to a vanilla profile')
@@ -57,15 +53,7 @@ export default function ProfileEditor() {
           toggleModActive(name)
         }}
       >
-        <div
-          className="cursor-pointer border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]"
-          style={{
-            backgroundColor: isHovered ? '#5A5B5C' : '#48494A',
-            padding: '1px',
-            paddingLeft: '4px',
-            paddingRight: '4px'
-          }}
-        >
+        <div className="cursor-pointer border-[3px] border-t-[#5a5b5c] border-l-[#5a5b5c] border-b-[#333334] border-r-[#333334] bg-[#48494a] p-[4px]">
           <p className="minecraft-seven text-white">{name}</p>
         </div>
       </div>
@@ -106,7 +94,7 @@ export default function ProfileEditor() {
     navigate('/profiles')
   }
 
-  const fetchMods = useCallback(() => {
+  useEffect(() => {
     const { mods } = GetMods()
     setAllMods(mods)
   }, [setAllMods])
@@ -115,36 +103,38 @@ export default function ProfileEditor() {
     loadProfile()
   }, [loadProfile])
 
-  useEffect(() => {
-    const intervalId = setInterval(fetchMods, 500) // Fetch every 5 seconds
-
-    return () => clearInterval(intervalId) // Cleanup interval on component unmount
-  }, [setAllMods, fetchMods])
-
   return (
     <MainPanel>
-      <div className="relative w-full h-full flex flex-col gap-[8px] border-[3px] border-[#1E1E1F] p-[8px] bg-[#48494A]">
+      <div className="w-full h-full flex flex-col border-[3px] border-[#1E1E1F] bg-[#48494A]">
         {/* Settings */}
-        <div>
-          <TextInput label="Profile Name" text={profileName} setText={setProfileName} />
-          <Dropdown
-            labelText="Minecraft Version"
-            value={profileMinecraftVersion}
-            setValue={setProfileMinecraftVersion}
-            // we don't support non-release versions right now so only show release lmao
-            options={allMinecraftVersions
-              .filter(ver => ver.versionType === MinecraftVersionType.Release)
-              .map(ver => ver.toString())}
-            id="minecraft-version"
-          />
-          <Dropdown
-            labelText="Runtime"
-            value={profileRuntime}
-            setValue={setProfileRuntime}
-            options={allRuntimes}
-            id="runtime-mod"
-          />
-          <TextInput label="Install Directory" text={profileInstallDir} setText={setProfileInstallDir} />
+        <div className="flex flex-col">
+          <div className="border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+            <TextInput label="Profile Name" text={profileName} setText={setProfileName} />
+          </div>
+          <div className="border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+            <Dropdown
+              labelText="Minecraft Version"
+              value={profileMinecraftVersion}
+              setValue={setProfileMinecraftVersion}
+              // we don't support non-release versions right now so only show release lmao
+              options={allMinecraftVersions
+                .filter(ver => ver.versionType === MinecraftVersionType.Release)
+                .map(ver => ver.toString())}
+              id="minecraft-version"
+            />
+          </div>
+          <div className="border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+            <Dropdown
+              labelText="Runtime"
+              value={profileRuntime}
+              setValue={setProfileRuntime}
+              options={allRuntimes}
+              id="runtime-mod"
+            />
+          </div>
+          <div className="border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px]">
+            <TextInput label="Install Directory" text={profileInstallDir} setText={setProfileInstallDir} />
+          </div>
         </div>
 
         {/* Mod Selection */}
@@ -153,8 +143,8 @@ export default function ProfileEditor() {
             <div className="h-full flex flex-col"></div>
           </div>
         ) : (
-          <div className="flex-grow flex justify-around gap-[8px]">
-            <div className=" w-[50%] h-full flex flex-col">
+          <div className="border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px] flex-grow flex justify-around gap-[8px]">
+            <div className="w-[50%] h-full flex flex-col">
               <p className="text-white minecraft-seven text-[14px]">Active Mods</p>
               <div className="border-[3px] border-[#1E1E1F] bg-[#313233] flex-grow">
                 {allMods.length > 0 ? (
@@ -182,7 +172,7 @@ export default function ProfileEditor() {
         )}
 
         {/* Profile Actions */}
-        <div className="flex justify-around gap-[8px]">
+        <div className="border-y-[3px] border-t-[#5a5b5c] border-b-[#333334] bg-[#48494a] p-[8px] flex justify-around gap-[8px]">
           <MinecraftButton text="Save Profile" onClick={() => saveProfile()} />
           <MinecraftButton text="Delete Profile" style={MinecraftButtonStyle.Warn} onClick={() => deleteProfile()} />
         </div>
